set delete success message and create update page

+++ src/Pages/MyReviews/MyReviews.js
@@ -2,6 +2,7 @@ import React, {useContext, useEffect, useState} from 'react';
 import {AuthContext} from '../../contexts/AuthProvider/AuthProvider';
 import SIngleReview from './SIngleReview';
 import {ScaleLoader} from 'react-spinners';
+import toast from 'react-hot-toast';
 const MyReviews = () => {
 	const {user, spinner, setSpinner} = useContext(AuthContext);
 	const [MyReviews, SetMyReviews] = useState();
@@ -15,17 +16,22 @@ const MyReviews = () => {
 			});
 	}, [setSpinner, user?.email]);
 	const handleDelete = (MyReview) => {
-		fetch(`http://localhost:5000/myreviews/${MyReview._id}`, {
-			method: 'DELETE',
-		})
-			.then((res) => res.json())
-			.then((data) => {
-				console.log(data);
-				if (data.deletedCount > 0) {
-					const restReviews = MyReviews.filter((rev) => rev._id !== MyReview._id);
-					SetMyReviews(restReviews);
-				}
-			});
+		const deleteRivew = window.confirm('Are You Sure To Delete Your Review');
+		if (deleteRivew) {
+			fetch(`http://localhost:5000/myreviews/${MyReview._id}`, {
+				method: 'DELETE',
+			})
+				.then((res) => res.json())
+				.then((data) => {
+					console.log(data);
+
+					if (data.deletedCount > 0) {
+						const restReviews = MyReviews.filter((rev) => rev._id !== MyReview._id);
+						SetMyReviews(restReviews);
+						toast.success('Delete Successfull');
+					}
+				});
+		}
 	};
 	return (
 		<>
